refactor: Reorganize Kafka email configuration and move email topic constants
