+++ src/main/java/com/shintadev/shop_dev_be/constant/KafkaConstants.java
@@ -3,12 +3,6 @@ package com.shintadev.shop_dev_be.constant;
 public record KafkaConstants() {
 
   /*
-   * Email Topics
-   */
-  public static final String VERIFICATION_EMAILS_TOPIC = "verification-emails";
-  public static final String PASSWORD_RESET_EMAILS_TOPIC = "password-reset-emails";
-  public static final String WELCOME_EMAILS_TOPIC = "welcome-emails";
-  /*
    * Email Groups
    */
   public static final String EMAIL_GROUP = "shop-dev-email-group";
