+++ src/main/java/com/shintadev/shop_dev_be/kafka/consumer/EmailConsumer.java
@@ -1,4 +1,4 @@
-package com.shintadev.shop_dev_be.kafka;
+package com.shintadev.shop_dev_be.kafka.consumer;
 
 import java.util.Map;
 
@@ -6,6 +6,7 @@ import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.stereotype.Component;
 
 import com.shintadev.shop_dev_be.constant.KafkaConstants;
+import com.shintadev.shop_dev_be.constant.kafka.KafkaTopic;
 import com.shintadev.shop_dev_be.service.common.EmailService;
 
 import lombok.RequiredArgsConstructor;
@@ -26,7 +27,7 @@ public class EmailConsumer {
    * 
    * @param emailData the email data
    */
-  @KafkaListener(topics = KafkaConstants.VERIFICATION_EMAILS_TOPIC, groupId = KafkaConstants.EMAIL_GROUP)
+  @KafkaListener(topics = KafkaTopic.VERIFICATION_EMAILS_TOPIC, groupId = KafkaConstants.EMAIL_GROUP)
   public void consumeVerificationEmail(Map<String, Object> emailData) {
     log.info("Consuming verification email for: {}", emailData.get(KafkaConstants.RECIPIENT_EMAIL_KEY));
     emailService.sendVerificationEmail(
@@ -41,7 +42,7 @@ public class EmailConsumer {
    * 
    * @param emailData the email data
    */
-  @KafkaListener(topics = KafkaConstants.PASSWORD_RESET_EMAILS_TOPIC, groupId = KafkaConstants.EMAIL_GROUP)
+  @KafkaListener(topics = KafkaTopic.PASSWORD_RESET_EMAILS_TOPIC, groupId = KafkaConstants.EMAIL_GROUP)
   public void consumePasswordResetEmail(Map<String, Object> emailData) {
     log.info("Consuming password reset email for: {}", emailData.get(KafkaConstants.RECIPIENT_EMAIL_KEY));
     emailService.sendPasswordResetEmail(
@@ -56,7 +57,7 @@ public class EmailConsumer {
    * 
    * @param emailData the email data
    */
-  @KafkaListener(topics = KafkaConstants.WELCOME_EMAILS_TOPIC, groupId = KafkaConstants.EMAIL_GROUP)
+  @KafkaListener(topics = KafkaTopic.WELCOME_EMAILS_TOPIC, groupId = KafkaConstants.EMAIL_GROUP)
   public void consumeWelcomeEmail(Map<String, Object> emailData) {
     log.info("Consuming welcome email for: {}", emailData.get(KafkaConstants.RECIPIENT_EMAIL_KEY));
     emailService.sendWelcomeEmail(
